Rename position-anchor:implicit to position-anchor:auto
See the spec discussion here:

https://github.com/w3c/csswg-drafts/issues/10312#issuecomment-2108836689

This is observable, but barely. After this CL, `implicit` won't be a
valid value for `position-anchor`, so it'll resolve to the default
value of `auto`. Meaning the behavior will be the same. But via the
CSSOM, getComputedStyle(element).positionAnchor will now return "auto"
instead of "implicit".

Fixed: 340206288
Change-Id: I75b05310638dca70aa6aa8807da8cfe7c9624a32