[document picture-in-picture] Propagate user activation to the opener
This CL changes browser-process-side user activation logic to allow
activations in document picture-in-picture windows to be used in their
opener windows, and activations in the opener windows to be used in
their document picture-in-picture windows.

This feature is behind an enabled-by-default feature flag.

Specification change:
https://github.com/WICG/document-picture-in-picture/pull/117

Chromestatus: https://chromestatus.com/feature/5185710702460928

Intent to Prototype:
https://groups.google.com/a/chromium.org/g/blink-dev/c/MwH0ODG4bec

Intent to Ship:
https://groups.google.com/a/chromium.org/g/blink-dev/c/AfzrMoCgccQ

Bug: 331246719
Change-Id: Iedad1e0cb60e04fc1ccf350405f04a399a2b72a1