Add allow unload=* to allowlist for frameset frames.
With unload deprecation, it becomes impossible to use unload in
`<frame>`s because there is no way to set an `allow` attribute on
them. This change adds an `unload=*` to their allowlist, meaning that
unload will be allowed if it's allowed in the parent and the child
sets the correct Permissions-Policy header.

Bug: 40286851
Change-Id: I593786ae88683986c740800848de7bfa6a364dd3